i have made change to home page which can be access when user details are filled
